pyproject.toml
[tool.poetry]
name = "plugin-resolver"
version = "0.1.0"
description = "OpenAPI resolver with support from Openai"
authors = ["Guilherme Ramos <guiramos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.4"
pydantic = "^1.10.7"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
